pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.9.2",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-silk>=5.4.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "celery>=5.5.3",
    "redis>=6.2.0"
]